TermsBuster/requirements.txt
altair==5.5.0
attrs==25.4.0
audioop-lts==0.2.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
filelock==3.20.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.10.1
networkx==3.5
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
protobuf==6.33.0
pyarrow==21.0.0
PyAudio==0.2.14
pycparser==2.23
pydeck==0.9.1
pypdfium2==5.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.7
watchdog==6.0.0
